working version with and without subject row
