+++ .travis.yml
@@ -35,7 +35,7 @@ jobs:
       script: docker run -v "$(pwd)":/mnt koalaman/shellcheck:stable src/main/server-certificate/*.sh
     - stage: integration
       name: "docker-maven-plugin integration tests"
-      script: mvn verify
+      script: mvn verify -Pdefault
     - stage: deploy
       name: "Deploy to DockerHub"
       script:
