--- luksi1.test.csr
@@ -1,17 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIICqDCCAZACAQAwYzELMAkGA1UEBhMCU0UxGDAWBgNVBAgMD1Zhc3RyYSBHb3Rh
-bGFuZDETMBEGA1UEBwwKR290aGVuYnVyZzEPMA0GA1UECgwGbHVrc2kxMRQwEgYD
-VQQDDAtsdWtzaTEudGVzdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEB
-AKOr97X6PcCqb6ajyHB5OxmaVkcQhMjeU+77FagipAUC9+mHSX+rhmhoPtL2s/pc
-LJxyfin/hImvlLl5MvBkLtxoYXBDHTkCXqKdlfJACjOmSi4oacfsgA+iRWGXQgkO
-/yyEKYu5NGCxVBZwAqZK1l3oyxIvTjMmEGclb92cQZjepBLLE6hS7f7iB25gyhN+
-HYsJnpEFYfH4ezbzp+c3QeHMjqcT+SJzW0kA5guWsWmYYL8ozvM7zzHTgiCYUDvc
-w2OvZJytd98ymAoKK7m0/Yw9weD87oXf184H2Tt3cww4sAryOCaNLFUiJvcNcgTb
-F7Q74U5yTBeMvRDQm6Wf4OkCAwEAAaAAMA0GCSqGSIb3DQEBCwUAA4IBAQBNcWC1
-JT1VmuUWQDG8TWc0n6PUONY4RA3iZKKK8rXj9Yz62Zv5t4O/5s5K4f95GyM01Jhy
-2oiGCvsWlbCTLR6TRNHplmhrxclfyLxvC9PZsZYCjh0LlgmxOGDZbNktr7KmG52q
-HULw6sQAxIWfmp2eV4XvZReWzJuY1EaPX9LoCAAiAHDiRhvDEXsT2dxZxz6/QKAL
-WCBj3i+5/t8hkE3gCE4C9VsV8YTwZsUpPqRp5NShS5P8kkV5kgzkhhMvnpyUEQ+m
-OFrZ8mSMTSOxhg2svUJqPuWxZtmvx+6FezphLUQrXIkw4St0rHvhK6lwzn6vD+a3
-nX87jPjMC849sCtV
------END CERTIFICATE REQUEST-----
