+++ src/main/root-certificate/docker-entrypoint.sh
@@ -13,4 +13,7 @@ export ROOT_CERT_EXPIRATION_DAYS
 
 /usr/local/bin/confd -onetime -backend env
 /usr/local/bin/create.certificates.sh
-echo "finished"
+
+if [[ $? -eq 0 ]]; then
+  echo "certificate created"
+fi
