+++ src/main/server-certificate/Dockerfile
@@ -14,7 +14,7 @@ RUN apk add --no-cache \
     go install github.com/kelseyhightower/confd && \
     rm -rf "/tmp/v${CONFD_VERSION}.tar.gz"
 
-FROM alpine:3.8
+FROM alpine:3.9
 
 COPY --from=confd /go/bin/confd /usr/local/bin/confd
 RUN chmod +x /usr/local/bin/confd
@@ -22,7 +22,8 @@ RUN chmod +x /usr/local/bin/confd
 COPY files/confd /etc/confd
 
 RUN apk update && \
-apk add --no-cache openssl=1.0.2q-r0 && \
+apk add --no-cache openssl=1.1.1a-r1 \
+sed=4.5-r0 && \
 rm -rf /var/cache/apk/*
 
 COPY docker-entrypoint.sh /docker-entrypoint.sh
