+++ src/main/server-certificate/docker-entrypoint.sh
@@ -1,11 +1,19 @@
 #!/bin/sh 
 
-SERVER_NAME=${SERVER_NAME:-"ci.test"}
-export SERVER_NAME
-SERVER_SUBJECT=${SERVER_SUBJECT:-"/C=SE/ST=Vastra Gotaland/L=Gothenburg/O=MyTest/CN=${SERVER_NAME}/"}
-export SERVER_SUBJECT
-# CRL_DISTRIBUTION_POINTS=${CRL_DISTRIBUTION_POINTS:-"URI:http://${SERVER_NAME}.crl.pem"}
-# export CRL_DISTRIBUTION_POINTS
+if [ "${SERVER_SUBJECT}x" != "x" ]; then
+  echo $SERVER_SUBJECT
+  SERVER_NAME=`echo $SERVER_SUBJECT | sed -r -n 's/.*CN=(.*)($|\/.*)/\1/p'`
+  echo "server subject"
+  echo $SERVER_NAME
+elif [ "${COMMON_NAME}x" != "x" ]; then
+  SERVER_NAME=${COMMON_NAME}
+  echo "common name"
+  echo $SERVER_NAME
+else
+  echo "You must indicate a SERVER_SUBJECT or a COMMON_NAME"
+  exit 1
+fi
+
 INTERMEDIATE_KEY_PASSWORD=${INTERMEDIATE_KEY_PASSWORD:-"Abcd1234"}
 export INTERMEDIATE_KEY_PASSWORD
 SERVER_KEY_PASSWORD=${SERVER_KEY_PASSWORD:-"Abcd1234"}
@@ -14,9 +22,12 @@ DEFAULT_CRL_DAYS=${DEFAULT_CRL_DAYS:-"1"}
 export DEFAULT_CRL_DAYS
 DEFAULT_DAYS=${DEFAULT_DAYS:-"1"}
 export DEFAULT_DAYS
-SERVER_CERT_EXPIRATION_DAYS=:${SERVER_CERT_EXPIRATION_DAYS:-"1"}
+SERVER_CERT_EXPIRATION_DAYS=${SERVER_CERT_EXPIRATION_DAYS:-"1"}
 export SERVER_CERT_EXPIRATION_DAYS
 
 /usr/local/bin/confd -onetime -backend env
 /usr/local/bin/create.certificates.sh
-echo "finished"
+
+if [[ $? -eq 0 ]]; then
+  echo "certificate created"
+fi
